Update to PoPS Core with explicit casts
This updates PoPS Core code to the latest version on main with explicit casts and conversions from floating point numbers to integers and booleans.

+++ inst/include/pest_host_table.hpp
@@ -103,7 +103,7 @@ public:
      * @param host Pointer to the host to get the information for
      * @return Mortality time lag value
      */
-    double mortality_time_lag(const HostPool* host) const
+    int mortality_time_lag(const HostPool* host) const
     {
         auto host_index = environment_.host_index(host);
         return mortality_time_lags_.at(host_index);
